Floating point exception solved. It was caused because I was dividing by zero in draw_line func
Need to add new file called rotate.c where theres a function to rotate every axis

The copy of the map you created must be modified along with the real mone when you translate and when you zoom. Keep that in mind for future bugs

Currently the rotation is done from the point (0, 0), which is a problem because the map escapes from the screen never to be seen again. Try to calculate the map center and then apply the difference after the rotation to accomplish the rotation from the center